Move meeting info retrieval to handler
